requirements.txt
git+ssh://git@github.com/revisto/instagrapi
python-telegram-bot
pika
redis